backend/requirements.txt
numpy==1.19.5
Flask==2.0.2
requests==2.25.1
gunicorn==20.1.0
tensorflow-cpu==2.5.0
Flask_RESTful==0.3.9
flask-cors==3.0.10